TASK: Prevent `neos/contentrepository-testsuite` to define the phpunit version
The package `neos/contentrepository-testsuite` is published on packagist to be installed as dev dependency for testing in libraries or projects.

It has helpers and step definitions for behat.

Phpunit is only used as dev dependency for tests in turn for that package.
And because we rely on `PHPUnit\Framework\Assert` for the step definitions.

The `Assert` api is pretty stable and available across php unit versions.

To allow using phpunit 11 and beyond in libraries or projects in combination with the testsuite we relax the constraint here. And we not only relax it to `>=9.6 <13` but we drop it because the composer manifest merger would otherwise put that into the distributions requirement which then might install a higher phpunit version for neos itself wich does not work.

Also both neos/contentrepository-structureadjustment and neos/contentrepository-nodemigration were turned into sugegestions

brianium/paratest is not used by the test suite and was moved to the behavior tests package instead.