BUGFIX: Avoid duplicating shadow nodes
This fixes a regression in handling shadow nodes
while publishing in nested workspaces.

Without this change we first create a new shadow
node to to replicating the move in the target
workspace and then try to adjust the existing
source shadow node to exist in the target,
which triggers a duplicate key error.
As commented, we can be sufficiently sure that the
move operation will take care of (not) creating
shadow nodes as necessary and any previously
existing shadow node in the source workspace can
be discarded safely.
The new "else" branch should probably not be
necessary, as if there was no move, there should
never be a shadow node to adjust. Just in case
though we leave this in place. Might clean up
some leftover shadow nodes.

Fixes: #5561
Related: #1608