TASK: Introduce dedicated and optional command "neos9preparation:preadjustfusion" instead of adding it as core migration for neos 8.4
While a core migration is certainly a more beautiful way it also comes with the dark magic of being just run without control whatsoever.

Doing the partial migration in 8.4 already is totally optional - this is not imminent when using core migrate.

And when applying the core migrations to an open source package it will become incompatible with earlier releases like Neos 8.3 which is not desired.