BUGFIX: Fix AFX AST positions to fix fusion migrations
positions in the AFX AST are now calculated properly:

```
{abc}
 1 3
```

`abc` starts at character `1` (`a`) and goes to character `3` (`c`)
position `0` and `4` are the outer braces.

This is also done correctly for spreads or when the expression is in a tag content
```
foo={abc}
     5 7

{...abc}
    4 6
```

The `EelExpressionTransformer` works with the positions differently and thus we offset them. But the semantic we now hold for the afx parser and the tests should ideally apply to all ast positions we use.