BUGFIX: Nodes show action does not handle `\JsonSerializable` objects and crashes if object is not StringAble
With https://github.com/neos/neos-ui/pull/3723 and https://github.com/neos/neos-ui/pull/3723 the `NodePropertyConverterService` returns `\JsonSerializable` objects directly in the expectation that they are serialized directly. This is also true for the usages within the Neos Ui.

But this legacy endpoint here in Neos doesnt do it. And throws

"Object of class Sitegeist\Archaeopteryx\Link could not be converted to string"

on some occasions whenever the endoint is used. Funnely the Neos Ui really does not care about the node properties rendered here via fluid in html so we could also return "nudelsuppe" here. Its just important that it not crashes.

https://github.com/neos/neos-development-collection/blob/1aeef6c02142b41e4626ea32434f06cd0f84e9bd/Neos.Neos/Resources/Private/Templates/Service/Nodes/Show.html#L30-L35

In the future we should have this whole thing killed :D